Pull request #4396: main: fix coverage for show snort cpu command
Merge in SNORT/snort3 from ~RISHACHO/snort3:cpu_usage_snort to master

Squashed commit of the following:

commit c8dcd376b5fe4a0e6a53eb31731badfe490885b8
Author: Rishabh Choudhary <rishacho@cisco.com>
Date:   Wed Jul 24 13:19:46 2024 +0530

    main: fix coverage
    Replace an impossible condition with assert statement.

+++ src/main/analyzer_command.h
@@ -218,9 +218,9 @@ public:
 
 private:
     int status = DAQ_SUCCESS;
-    float cpu_usage_30s = 0.0;
-    float cpu_usage_120s = 0.0;
-    float cpu_usage_300s = 0.0;
+    double cpu_usage_30s = 0.0;
+    double cpu_usage_120s = 0.0;
+    double cpu_usage_300s = 0.0;
     int instance_num = 0;
     std::mutex cpu_usage_mutex;
 };
